requirements.txt
selenium
fpdf
pillow
webdriver-manager